requirements.txt
aiohttp==3.12.12
aiofiles==23.2.1
supabase==2.15.2
python-dotenv==1.0.0
psutil==5.9.8
httpx==0.26.0
backoff==2.2.1
tenacity==8.2.3